+++ tests/language_models/test_rts.py
@@ -1,5 +1,6 @@
 import pytest
 import torch
+from pytorch_lightning import seed_everything
 from transformers import BertTokenizer
 
 from transformers_lightning.language_modeling import IGNORE_IDX
@@ -29,8 +30,7 @@ rts = RandomTokenSubstitution(tok, whole_word_swapping=True)
 )
 def test_language_model(seed, sentence, masking, new_ids):
 
-    torch.random.manual_seed(seed)
-    torch.cuda.random.manual_seed(seed)
+    seed_everything(seed)
 
     input_ids = torch.tensor([tok.encode(sentence)])
     words_tails_mask = whole_word_tails_mask(input_ids, tok)