+++ transformers_lightning/datamodules/adapter_datamodule.py
@@ -1,4 +1,5 @@
 from argparse import Namespace
+from typing import Union
 
 from transformers_lightning.adapters.super_adapter import SuperAdapter
 from transformers_lightning.datamodules.super_datamodule import SuperDataModule
@@ -7,7 +8,7 @@ from transformers_lightning.datasets.map_dataset import MapDataset
 
 
 class AdaptersDataModule(SuperDataModule):
-    """
+    r"""
     AdaptersDataModule should be used when you want to read and tokenizer data on-the-fly.
     It implements some simple methods to check whether training, val or testing is required.
     It work with adapters: you could define them inside the `__init__()` method or pass them
@@ -50,7 +51,8 @@ class AdaptersDataModule(SuperDataModule):
                 predict_adapter, SuperAdapter
             ), "Argument `predict_adapter` must be of type `SuperAdapter`"
             self.predict_adapter = predict_adapter
-        """
+
+        r"""
         This space should be used to instantiate the Adapters it they were not passed through the kwargs
 
         >>> self.train_adapter = CSVAdapter(self.hyperparameters, "pre-training/train.tsv", delimiter="\t")
@@ -59,31 +61,37 @@ class AdaptersDataModule(SuperDataModule):
         >>> self.predict_adapter = CSVAdapter(self.hyperparameters, "pre-training/predict.tsv", delimiter="\t")
         """
 
+    def get_dataset(self, adapter: SuperAdapter) -> Union[MapDataset, TransformersIterableDataset]:
+        r""" Return iterable or map dataset from adapter. """
+        if self.hyperparameters.iterable:
+            return TransformersIterableDataset(self.hyperparameters, adapter, self.trainer)
+        else:
+            return MapDataset(self.hyperparameters, adapter, self.trainer)
+
     # Optional, called for every GPU/machine (assigning state is OK)
     def setup(self, stage=None):
-        """
+        r"""
         Load datasets only if respective Adapter are defined.
         This implementation should be enough for most subclasses.
         """
-        dataset_class = MapDataset if not self.hyperparameters.iterable else TransformersIterableDataset
 
         if stage == 'fit':
             if self.train_adapter is not None:
-                self.train_dataset = dataset_class(self.hyperparameters, self.train_adapter, self.trainer)
+                self.train_dataset = self.get_dataset(self.train_adapter)
             if self.valid_adapter is not None:
-                self.valid_dataset = dataset_class(self.hyperparameters, self.valid_adapter, self.trainer)
+                self.valid_dataset = self.get_dataset(self.valid_adapter)
 
         elif stage == 'test':
             if self.test_adapter is not None:
                 if isinstance(self.test_adapter, SuperAdapter):
                     self.test_adapter = [self.test_adapter]
                 self.test_dataset = [
-                    dataset_class(self.hyperparameters, adapter, self.trainer) for adapter in self.test_adapter
+                    self.get_dataset(adapter) for adapter in self.test_adapter
                 ]
 
         elif stage == 'predict':
             if self.predict_dataset is not None:
-                self.predict_adapter = dataset_class(self.hyperparameters, self.predict_adapter, self.trainer)
+                self.predict_adapter = self.get_dataset(self.predict_adapter)
 
     def do_train(self):
         return self.train_adapter is not None